refactor: factorize code turning a tag name into a version
