+++ src/bin/clydetools/main.rs
@@ -13,6 +13,7 @@ pub mod check_package;
 pub mod fetch;
 pub mod github_fetcher;
 pub mod gitlab_fetcher;
+pub mod version_utils;
 
 #[macro_use]
 extern crate lazy_static;
